Add prompt guards to prevent unintended overwrites

+++ index.js
@@ -3,13 +3,15 @@
 const fs = require('fs-extra')
 const { spawn } = require('child_process')
 const package = require('./template/package')
+const prompt = require('prompt')
+const prompts = require('./prompts')
 
 function performYarnInstall() {
   const child = spawn('yarn')
   child.stdout.on('data', (data) => console.log(`${data}`))
   child.stderr.on('data', (data) => console.error(`${data}`))
   child.on('exit', () => {
-    notifyYarnSuccess()
+    prompts.notifyYarnSuccess()
   })
 }
 
@@ -17,64 +19,64 @@ if (!fs.existsSync('.git')) {
 
   // If dir is not under VC, just do a straight copy
   fs.copy(process.mainModule.filename.replace('index.js', '') + 'template', process.env.PWD + '/')
-    .catch(err => console.error(err))  
+    .catch(err => console.error(err))
     .then(() => {
-      notifyCopySuccess()
+      prompts.notifyCopySuccess()
       performYarnInstall()
     })
 
 } else {
 
-  notifyPackageMerge()
+  prompt.start()
+  prompt.get(prompts.existingRepo, function (error, result) {
 
-  const dependencies = package.dependencies
-  const devDependencies = package.devDependencies
-  let newPackage = {}
+    if (!error && result && (/y/i).test(result.confirm)) {
 
-  if (fs.existsSync('package.json')) {
-    newPackage = JSON.parse(fs.readFileSync('package.json', 'utf-8'))
-  } else {
-    newPackage = {}
-  }
+      if (!fs.existsSync('package.json')) {
 
-  // Merge package.jsons
-  Object.keys(package).map(key => {
-    if (typeof package[key] === 'object') {
-      newPackage[key] = Object.assign({}, newPackage[key], package[key])
+        prompt.get(prompts.noPackageJSON, function (error, result) {
+          if (!error && result && (/y/i).test(result.confirm)) {
+
+          } else {
+            process.exit()
+          }
+        })
+
+      }
+
+      const dependencies = package.dependencies
+      const devDependencies = package.devDependencies
+      let newPackage = {}
+
+      if (fs.existsSync('package.json')) {
+        newPackage = JSON.parse(fs.readFileSync('package.json', 'utf-8'))
+      } else {
+        newPackage = {}
+      }
+
+      // Merge package.jsons
+      Object.keys(package).map(key => {
+        if (typeof package[key] === 'object') {
+          newPackage[key] = Object.assign({}, newPackage[key], package[key])
+        }
+      })
+
+      // Copy template over
+      fs.copy(process.mainModule.filename.replace('index.js', '') + 'template', process.env.PWD + '/')
+        .catch(err => console.error(err))
+        .then(() => {
+
+          prompts.notifyCopySuccess()
+
+          // Overwrite package.json, retaining previous data
+          fs.writeFileSync('package.json', JSON.stringify(newPackage))
+
+          performYarnInstall()
+
+        })
+
+    } else {
+      process.exit()
     }
   })
-
-  // Copy template over
-  fs.copy(process.mainModule.filename.replace('index.js', '') + 'template', process.env.PWD + '/')
-    .catch(err => console.error(err))
-    .then(() => {
-
-      notifyCopySuccess()
-
-      // Overwrite package.json, retaining previous data
-      fs.writeFileSync('package.json', JSON.stringify(newPackage))
-
-      performYarnInstall()
-
-    })
-
-}
-
-function notifyPackageMerge() {
-  console.log('')
-  console.log('------------------------ WARNING -------------------------')
-  console.log('------- Current directory is under version control -------')
-  console.log('------------ Your package.json will be merged ------------')
-  console.log('----------- Anything else will be overwritten ------------')
-  console.log('------------------------ WARNING -------------------------')
-  console.log('')
-}
-function notifyCopySuccess() {
-  console.log('')
-  console.log('---------- Template copied to current directory ----------')
-  console.log('')
-}
-function notifyYarnSuccess() {
-  console.log('----------------- Yarn install complete ------------------')
-  console.log('')
 }
